Move toggleCheckbox() to the right place and trigger it in jQuery style
