+++ caravel/templates/caravel/base.html
@@ -10,3 +10,8 @@
     {{super()}}
     <script src="/static/assets/javascripts/dist/css-theme.entry.js"></script>
   {% endblock %}
+
+  {% block tail_js %}
+      {{super()}}
+      <script src="/static/assets/javascripts/dist/common.entry.js"></script>
+  {% endblock %}
